Updated SATSolverTest.Java
Included the clauses into the formula and added the SATSolver enabler part. The comments can be used to run the print to check. I removed the literals and its negation from each line since (a or -a) is always true so we can disregard it from the function. Requires SATSolver to be finished first to run properly, if not only the printed clauses in formula can be seen. Also, please input a working directory into argument or nothing will appear.